# Chore: lazy loading for rules