--- a/lib/cli.js
+++ b/lib/cli.js
@@ -81,23 +81,28 @@ function translateOptions(cliOptions) {
  */
 function printResults(engine, results, format, outputFile) {
     let formatter;
-    let rules;
+    let rulesMeta;
 
     try {
         formatter = engine.getFormatter(format);
-        rules = engine.getRules();
     } catch (e) {
         log.error(e.message);
         return false;
     }
 
-    const rulesMeta = {};
+    const output = formatter(results, {
+        get rulesMeta() {
+            if (!rulesMeta) {
+                const rules = engine.getRules();
 
-    rules.forEach((rule, ruleId) => {
-        rulesMeta[ruleId] = rule.meta;
+                rulesMeta = {};
+                for (const [ruleId, rule] of rules) {
+                    rulesMeta[ruleId] = rule.meta;
+                }
+            }
+            return rulesMeta;
+        }
     });
-
-    const output = formatter(results, { rulesMeta });
 
     if (output) {
         if (outputFile) {