--- a/lib/cli.js
+++ b/lib/cli.js
@@ -93,10 +93,8 @@ function printResults(engine, results, format, outputFile) {
     const output = formatter(results, {
         get rulesMeta() {
             if (!rulesMeta) {
-                const rules = engine.getRules();
-
                 rulesMeta = {};
-                for (const [ruleId, rule] of rules) {
+                for (const [ruleId, rule] of engine.getRules()) {
                     rulesMeta[ruleId] = rule.meta;
                 }
             }