--- a/lib/testers/rule-tester.js
+++ b/lib/testers/rule-tester.js
@@ -307,8 +307,6 @@ class RuleTester {
 
         linter.defineRules(this.rules);
 
-        const ruleMap = linter.getRules();
-
         /**
          * Run the rule for the given item
          * @param {string|Object} item Item to run the rule against
@@ -398,7 +396,7 @@ class RuleTester {
                 }
             }
 
-            validate(config, "rule-tester", ruleMap.get.bind(ruleMap));
+            validate(config, "rule-tester", id => (id === ruleName ? rule : null));
 
             return {
                 messages: linter.verify(code, config, filename, true),